schmalz: texture is no at the state of a dummy transform: - it is linear - it has a solver - but it is quite simple (and of course not the correct transform) - but it provides the main interfaces
To create the solver I had to split the code in two parts:
There is now a directory trunk/texture with the library of the transform function.
An example with test code is in trunk/examples/texture.
I was not able to avoid this splitting because otherwise I had a lot of trouble with the dependencies.

I also added the svn:ignore property to a lot of directories to make svn ignore the files created by autoconf, automake, gcc, ... etc.


git-svn-id: https://euler.mathematik.tu-chemnitz.de/repos/nfft/trunk@272 c8b3e4ef-9df4-0310-9d76-b6dd82e1323a