Feature/octave tests (#65)
* ported CUnit-like tests for nfftmex

* add output of failed tests

* use sudo and newer octave for matlab interface tests

* always show cpuinfo

* slightly increased test threshold for bessel_i0 in single precision:
'random' failures of make check occurred on travis-ci when using clang 5.0.0 and single precision. Only one single check slightly failed:
i0[ 1.50000000E+01] =  3.39648438E+05 err_rel =  2.76019932E-06 >=  1.90734863E-06 -> FAIL
This check only seems to fail when the used cpufamily/model is 6/63 on but pass for 6/45 and 6/62. Presumably, the main difference is that 6/63 supports avx2. On a computer with clang 3.7.0 and cpu 6/94 with avx2, the identical error value 2.76019932E-06 is obtained.
When using gcc, the error value is just slightly below the threshold.
I have increased the threshold to 2.861E-06 (24*eps instead of 16*eps) for now.

* use MATLAB / Octave path from configure script

* change matlab call compatible to windows and macos

* add ABS_SRCDIR and more error handling to matlab/octave tests